Fixed default values for properties
