+++ src/FluentSetups.UnitTests/CodeGenerationTests/SetupMemberTests.cs
@@ -298,75 +298,7 @@ public class SetupMemberTests
       result.Print();
    }
    
-   [TestMethod]
-   public void EnsureSetupNameIsGeneratedCorrectlyForFieldsWithDefaultValues()
-   {
-      var code = @"namespace RonnyTheRobber
-                   {   
-                      using FluentSetups;
 
-                      internal class Person 
-                      {  
-                         public string Name{ get; set; }
-                      }
-
-                      [FluentSetup(typeof(Person))]
-                      internal partial class PersonSetup
-                      {
-                          [FluentMember]
-                          private string name = ""Robert"";
-                      }
-                   }";
-
-      var result = Setup.SourceGeneratorTest()
-         .WithSource(code)
-         .Done();
-
-      result.Should().NotHaveWarnings().And
-         .HaveClass("RonnyTheRobber.PersonSetup")
-         .WhereMethod("SetupName")
-         .IsProtected()
-         .Contains("target.Name = name;")
-         .NotContains("nameWasSet");
-         
-
-      result.Print();
-   }
-
-   [TestMethod]
-   public void EnsureSetupNameIsGeneratedCorrectlyForPropertiesWithDefaultValues()
-   {
-      var code = @"namespace RonnyTheRobber
-                   {   
-                      using FluentSetups;
-
-                      internal class Person 
-                      {  
-                         public string Name { get; set; }
-                      }
-
-                      [FluentSetup(typeof(Person))]
-                      internal partial class PersonSetup
-                      {
-                          [FluentMember]
-                          private string Name { get; set; } = ""Robert"";
-                      }
-                   }";
-
-      var result = Setup.SourceGeneratorTest()
-         .WithSource(code)
-         .Done();
-
-      result.Should().NotHaveWarnings().And
-         .HaveClass("RonnyTheRobber.PersonSetup")
-         .WhereMethod("SetupName")
-         .IsProtected()
-         .Contains("target.Name = Name;")
-         .NotContains("nameWasSet");
-         
-
-      result.Print();
-   }
 
    #endregion
 }
